improve structure in types crate
